refactor(tests): move mock models to tests/mocks directory
